requirements.txt
altgraph==0.17.4
macholib==1.16.3
packaging==25.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyserial==3.5
websockets==15.0.1
